Fix build

+++ http/SSLSocket.h
@@ -22,8 +22,8 @@ public:
 
 	virtual int Connect(const struct sockaddr *address, socklen_t addrLen);
 
-	virtual int Read(void* data, const size_t& length);
-	virtual int Write(const void* data, const size_t& length);
+	virtual size_t Read(void* data, const size_t& length);
+	virtual size_t Write(const void* data, const size_t& length);
 
 private:
 	void _SSLInit();
